requirements.txt
python-telegram-bot==21.5
anthropic>=0.28
feedparser==6.0.11
httpx==0.27.0
APScheduler==3.10.4
